pyproject.toml
[tool.poetry]
name = "snp-project"
version = "0.1.0"
description = ""
authors = ["akelaPro <ilya_chernov_2022@mail.ru>"]
readme = "README.md"
packages = [
    { include = "snpProject" }
]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
psycopg2 = "^2.9"
python-decouple = "^3.5"
dj-database-url = "^0.5.0"
pillow = "^9.5"
ipython = "^8.10"
django-extensions = "^3.2"
django-imagekit = "^5.0"
djangorestframework = "^3.15.2"
djoser = "^2.3.1"
djangorestframework-simplejwt = "^5.3.1"
channels = "^4.2.0"
d = "^0.2.2"
channels-redis = "^4.2.1"
daphne = "^4.1.2"
celery = "^5.4.0"
whitenoise = "^6.8.2"
drf-spectacular = "^0.28.0"
django-service-objects = "^0.7.1"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
